fix: the rest of the review findings, and the gaps that let them through
Ten findings and a cut list. The ones that changed a payload:

A comment could no longer say it was a reply or which thread it belonged
to -- Parent and Reply were dropped when the model replaced the raw
object, so a caller holding a reply id had no way back to the root that
resolve, reopen and react take.

--describe answered at two levels. A derived schema describes the data
payload; the three hand-written bulk schemas describe the whole
envelope, and both were served under one field, so a consumer validating
envelope.data passed for a declared command and rejected every bulk
document. They now answer at the payload level, without the $id that
pointed into a directory this project no longer publishes. The bulk
status enum also regained cancelled, which the workflow has emitted
since the run could be interrupted.

`pr comment list --full` published every comment on the pull request
beside a state field it had ignored. It now carries the comments the
thread filters kept, which is what --state and --tasks-only were asked
for.

`repo ssh-key list --json` had lost meta.limitReached, so a truncated
listing looked complete.

`build stats` looked its tallies up by the raw argument while the
service keyed them by the trimmed one, so a padded commit id reported
zeros. `build required create` and `update` rendered an id and a key
list, dropping the four fields that say what the check enforces.

RestInsightReport.createdDate decoded through float32, whose mantissa
cannot hold an epoch millisecond: every report timestamp was rounded to
the nearest few minutes. The sanitizer already widens createdDate where
the spec declares it as a date-time string; it now widens a bare number
for the same reason.

Participant advertised a display name the endpoint cannot supply.

Then the gaps. --describe on a group, on help, or on completion printed
human text and exited 0 -- a success a caller cannot parse. Cobra's two
commands are now installed before the walk, so they answer like
everything else, and TestEveryCommandIsModelled caught all five the
moment they entered the tree. A group says it is a group. The exemption
maps went unchecked, so an entry could name a renamed command or shadow
a declaration it should have lost to; both are now tested. Cobra's
entries do not count against the exemption budget, which is about
commands bb wrote.

The docs said version v2, meta.contract, exit_code, credential_storage
and review_summary.counts_source, none of which exist, and told readers
most commands publish no schema, which is no longer true.

+++ internal/cli/cmd/insights/result.go
@@ -26,7 +26,7 @@ type Report struct {
 	Reporter    string        `json:"reporter,omitempty" jsonschema:"Which system produced the report."`
 	Link        string        `json:"link,omitempty" jsonschema:"Link back to the reporting system."`
 	LogoURL     string        `json:"logoUrl,omitempty" jsonschema:"Logo shown beside the report in the Bitbucket UI."`
-	CreatedDate float64       `json:"createdDate,omitempty" jsonschema:"When the report was created, in milliseconds since the epoch."`
+	CreatedDate int64         `json:"createdDate,omitempty" jsonschema:"When the report was created, in milliseconds since the epoch."`
 	Data        []ReportDatum `json:"data,omitempty" jsonschema:"Labelled figures attached to the report."`
 }
 
@@ -99,7 +99,7 @@ func reportFrom(upstream openapigenerated.RestInsightReport) Report {
 		converted.Result = string(*upstream.Result)
 	}
 	if upstream.CreatedDate != nil {
-		converted.CreatedDate = float64(*upstream.CreatedDate)
+		converted.CreatedDate = *upstream.CreatedDate
 	}
 	if upstream.Data != nil {
 		converted.Data = make([]ReportDatum, 0, len(*upstream.Data))
